Extract shared ValidationBadge component
